Build the Acme tools. "acd" builds now but I don't have a machine with a cdrom drive to try it.
build.json files now include /sys/src/cmd/cmd.json explicitly.

(I'm a big idiot and committed some binaries by mistake last try)

Last update: add acme to the top-level build.json and add acd
to /acme/bin/source/build.json since it works.

+++ acme/wiki/src/awiki.h
@@ -104,7 +104,7 @@ struct Wiki {
 	int n;
 	int dead;
 	Window *win;
-	ulong time;
+	uint32_t time;
 	int linked;
 	Wiki *next;
 	Wiki *prev;
